Update activity service join & leave

+++ backend/activity-service/src/app.service.ts
@@ -4,18 +4,23 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Activity } from 'src/entities/Activity';
+import { Activity } from 'src/entities/activity.entity';
 import { Repository } from 'typeorm';
 import {
   CreateActivityRequestDto,
+  DeleteActivityRequestDto,
+  JoinActivityRequestDto,
   UpdateActivityRequestDto,
 } from './activity.dto';
+import { ActivityParticipant } from './entities/activity-participant.entity';
 
 @Injectable()
 export class AppService {
   constructor(
     @InjectRepository(Activity)
     private activitiesRepository: Repository<Activity>,
+    @InjectRepository(ActivityParticipant)
+    private activityParticipantsRepository: Repository<ActivityParticipant>,
   ) {}
 
   async findActivity(id: string): Promise<Activity> {
@@ -31,15 +36,36 @@ export class AppService {
     const activities = await this.activitiesRepository.find();
     console.log('activities', activities);
     return {
-      activities: activities.map((activity) => ({
-        id: activity.id,
-        title: activity.title,
-        description: activity.description,
-        startDate: activity.startDate,
-        endDate: activity.endDate,
-        ownerId: activity.ownerId,
-      })),
+      activities: activities,
     };
+  }
+
+  async joinActivity(data: JoinActivityRequestDto): Promise<Activity> {
+    const activity = await this.findActivity(data.id);
+    if (!activity || activity.status === 'CLOSED') {
+      throw new NotFoundException('Activity is already closed or not found');
+    }
+    activity.currentParticipants += 1;
+    const participant = await this.activityParticipantsRepository.save({
+      activityId: data.id,
+      userId: data.userId,
+    });
+    const res = await this.activitiesRepository.save(activity);
+    return res;
+  }
+
+  async leaveActivity(data: JoinActivityRequestDto): Promise<Activity> {
+    const activity = await this.findActivity(data.id);
+    if (!activity || activity.status === 'CLOSED') {
+      throw new NotFoundException('Activity is already closed or not found');
+    }
+    activity.currentParticipants -= 1;
+    await this.activityParticipantsRepository.delete({
+      userId: data.userId,
+      activityId: data.id,
+    });
+    const res = await this.activitiesRepository.save(activity);
+    return res;
   }
 
   async createActivity(
@@ -51,19 +77,19 @@ export class AppService {
       startDate: new Date(createActivityDto.startDate),
       endDate: new Date(createActivityDto.endDate),
     });
-    activity.ownerId = createActivityDto.ownerId;
-    return this.activitiesRepository.save(activity);
+    const res = await this.activitiesRepository.save(activity);
+    return res;
   }
 
   async updateActivity(
     updateActivityDto: UpdateActivityRequestDto,
   ): Promise<Activity> {
     const activity = await this.findActivity(updateActivityDto.id);
-    // if (activity.ownerId !== updateActivityDto.user_id) {
-    //   throw new UnauthorizedException(
-    //     'You are not authorized to update this activity',
-    //   );
-    // }
+    if (activity.ownerId !== updateActivityDto.ownerId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this activity',
+      );
+    }
     if (updateActivityDto.title) activity.title = updateActivityDto.title;
     if (updateActivityDto.description)
       activity.description = updateActivityDto.description;
@@ -74,11 +100,19 @@ export class AppService {
     return this.activitiesRepository.save(activity);
   }
 
-  async deleteActivity(id: string): Promise<void> {
-    const result = await this.activitiesRepository.delete({ id });
-    if (result.affected === 0) {
-      throw new NotFoundException(`Activity with ID ${id} not found`);
+  async deleteActivity(data: DeleteActivityRequestDto): Promise<boolean> {
+    const { id, ownerId } = data;
+    const activity = await this.findActivity(id);
+    if (
+      !activity ||
+      activity.status === 'CLOSED' ||
+      activity.ownerId !== ownerId
+    ) {
+      throw new UnauthorizedException(
+        'You are not authorized to delete this activity or the activity is already closed',
+      );
     }
-    return;
+    const result = await this.activitiesRepository.delete({ id });
+    return true;
   }
 }
